Made Navboi component component and chucked it in the home component
